Cleaning up Main and misspellings/typos
Added command to delete empty space in flags.
I assume that the line l = len(flags) -1 was to make sure that the empty space that appears when there's a space between the last input and < it wouldn't cause an error. This would make it so that the loop would not read the last item if that space wasn't there/forgotten. The command to delete empty spaces in flags should resolve this issue so l = len(flags).
Made changes to algorithm flag while loop so the check to see if next item is a flag rather than an algorithm into elif statement for readability. Also made it so that once it checks if next value is not the end, i iterates to next value. Code changed to reflect this new behavior.
Added iscmdInvalid check to while loop so loop stops when command is invalid.
Added command: if no algorithm flag was provided, then add all algorithms from project 1 instructions: fcfs, sjf, srtn, rr 10, rr 50, and rr 100

Fixed some misspellings and typos.